+++ modules/api-svc/src/index.js
@@ -27,6 +27,8 @@
 'use strict';
 
 const EventEmitter = require('node:events');
+const http = require('http');
+const https = require('https');
 const _ = require('lodash');
 const { name, version } = require('../../../package.json');
 
@@ -76,6 +78,9 @@ class Server extends EventEmitter {
             logger: this.logger
         });
 
+        // Create shared Mojaloop agents for switch communication (used by both servers)
+        this.mojaloopSharedAgents = this._createMojaloopSharedAgents(this.conf);
+
         this.wso2 = createAuthClient(conf, logger);
         this.wso2.auth.on('error', (msg) => {
             this.emit('error', 'WSO2 auth error in InboundApi', msg);
@@ -87,6 +92,7 @@ class Server extends EventEmitter {
             this.cache,
             this.metricsClient,
             this.wso2,
+            this.mojaloopSharedAgents,
         );
         this.inboundServer.on('error', (...args) => {
             this.logger.isErrorEnabled && this.logger.push({ args }).error('Unhandled error in Inbound Server');
@@ -99,6 +105,7 @@ class Server extends EventEmitter {
             this.cache,
             this.metricsClient,
             this.wso2,
+            this.mojaloopSharedAgents,
         );
         this.outboundServer.on('error', (...args) => {
             this.logger.isErrorEnabled && this.logger.push({ args }).error('Unhandled error in Outbound Server');
@@ -288,12 +295,15 @@ class Server extends EventEmitter {
             // eslint-disable-next-line no-console
             console.time(stopStartLabel);
             await this.inboundServer.stop();
+
+            this.mojaloopSharedAgents = this._createMojaloopSharedAgents(newConf);
             this.inboundServer = new InboundServer(
                 newConf,
                 this.logger,
                 this.cache,
                 this.metricsClient,
                 this.wso2,
+                this.mojaloopSharedAgents,
             );
             this.inboundServer.on('error', (...args) => {
                 const errMessage = 'Unhandled error in Inbound Server';
@@ -313,12 +323,15 @@ class Server extends EventEmitter {
             // eslint-disable-next-line no-console
             console.time(stopStartLabel);
             await this.outboundServer.stop();
+
+            this.mojaloopSharedAgents = this._createMojaloopSharedAgents(newConf);
             this.outboundServer = new OutboundServer(
                 newConf,
                 this.logger,
                 this.cache,
                 this.metricsClient,
                 this.wso2,
+                this.mojaloopSharedAgents,
             );
             this.outboundServer.on('error', (...args) => {
                 const errMessage = 'Unhandled error in Outbound Server';
@@ -448,6 +461,37 @@ class Server extends EventEmitter {
             this.backendEventHandler?.stop(),
             this.fspiopEventHandler?.stop(),
         ]);
+    }
+
+    _createMojaloopSharedAgents(conf) {
+        const httpAgent = new http.Agent({
+            keepAlive: true,
+            maxSockets: conf.outbound?.maxSockets || 256,
+        });
+
+        // Create HTTPS agent based on TLS configuration for Mojaloop switch communication
+        const httpsAgentOptions = {
+            keepAlive: true,
+            maxSockets: conf.outbound?.maxSockets || 256,
+        };
+
+        // Apply TLS configuration if mTLS is enabled for switch communication
+        if (conf.outbound?.tls?.mutualTLS?.enabled && conf.outbound?.tls?.creds) {
+            Object.assign(httpsAgentOptions, conf.outbound.tls.creds);
+        }
+
+        const httpsAgent = new https.Agent(httpsAgentOptions);
+
+        // Prevent accidental logging of agent internals
+        httpAgent.toJSON = () => ({ type: 'HttpAgent', keepAlive: httpAgent.keepAlive });
+        httpsAgent.toJSON = () => ({ type: 'HttpsAgent', keepAlive: httpsAgent.keepAlive });
+
+        this.logger.isInfoEnabled && this.logger.info('Created shared HTTP and HTTPS agents for Mojaloop switch communication');
+
+        return {
+            httpAgent,
+            httpsAgent
+        };
     }
 }